requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
psycopg2-binary==2.9.5
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-JWT-Extended==4.4.4
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
python-dotenv==0.21.1
Werkzeug==2.2.2